pyproject.toml
# SPDX-FileCopyrightText: 2025 LichtFeld Studio Authors
# SPDX-License-Identifier: GPL-3.0-or-later

[project]
name = "densification"
version = "0.5.0"
description = "Dense point cloud initialization using RoMa v2 matching"
author = "Shady Gmira"
min_lichtfeld_version = "0.5.0"


dependencies = [
    "torch",
    "torchvision",
    "numpy==2.4.1",
    "pycolmap",
    "Pillow",
    "scipy",
    "tqdm",
    "einops>=0.8.1",
    "rich>=14.2.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[tool.lichtfeld]
hot_reload = true
